+++ .idea/workspace.xml
@@ -6,15 +6,21 @@
   <component name="ChangeListManager">
     <list default="true" id="567eefc8-ab3b-4aca-8c19-004a4a7975a5" name="Changes" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/Lab02/Server/Server.go" beforeDir="false" afterPath="$PROJECT_DIR$/Lab02/Server/Server.go" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/Lab02_Phuc/Server/Server.go" beforeDir="false" afterPath="$PROJECT_DIR$/Lab02_Phuc/Server/Server.go" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Lab05/Client/Client.go" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Lab05/README.md" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Lab05/Server/Server.go" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Lab05/Shared/consistent_hash.go" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Lab05/Shared/consistent_hash_test.go" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Lab05/Shared/types.go" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Lab05/go.mod" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Lab05/integration_test.go" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
     <option name="LAST_RESOLUTION" value="IGNORE" />
   </component>
-  <component name="GOROOT" url="file:///usr/local/go" />
+  <component name="GOROOT" url="file://$USER_HOME$/go/pkg/mod/golang.org/toolchain@v0.0.1-go1.23.6.windows-amd64" />
   <component name="Git.Settings">
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
   </component>
@@ -43,16 +49,17 @@
   &quot;keyToString&quot;: {
     &quot;ASKED_SHARE_PROJECT_CONFIGURATION_FILES&quot;: &quot;true&quot;,
     &quot;Go Build.go build Server.go.executor&quot;: &quot;Run&quot;,
+    &quot;Go Test.integration_test.go.executor&quot;: &quot;Run&quot;,
     &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
     &quot;RunOnceActivity.git.unshallow&quot;: &quot;true&quot;,
     &quot;RunOnceActivity.go.formatter.settings.were.checked&quot;: &quot;true&quot;,
     &quot;RunOnceActivity.go.migrated.go.modules.settings&quot;: &quot;true&quot;,
     &quot;RunOnceActivity.go.modules.go.list.on.any.changes.was.set&quot;: &quot;true&quot;,
-    &quot;git-widget-placeholder&quot;: &quot;Phuc&quot;,
+    &quot;git-widget-placeholder&quot;: &quot;Huan&quot;,
     &quot;go.import.settings.migrated&quot;: &quot;true&quot;,
     &quot;go.sdk.automatically.set&quot;: &quot;true&quot;,
     &quot;ignore.virus.scanning.warn.message&quot;: &quot;true&quot;,
-    &quot;last_opened_file_path&quot;: &quot;/Users/phucnguyen/Desktop/Master/HP1-K34/Distributed systems/Golang/Distributed_System&quot;,
+    &quot;last_opened_file_path&quot;: &quot;C:/Users/PC&quot;,
     &quot;node.js.detected.package.eslint&quot;: &quot;true&quot;,
     &quot;node.js.selected.package.eslint&quot;: &quot;(autodetect)&quot;,
     &quot;nodejs_package_manager_path&quot;: &quot;npm&quot;
@@ -63,7 +70,7 @@
       <recent name="$PROJECT_DIR$" />
     </key>
   </component>
-  <component name="RunManager">
+  <component name="RunManager" selected="Go Test.integration_test.go (1)">
     <configuration name="go build Server.go" type="GoApplicationRunConfiguration" factoryName="Go Application" temporary="true" nameIsGenerated="true">
       <module name="Distributed_System" />
       <working_directory value="$PROJECT_DIR$" />
@@ -72,8 +79,28 @@
       <filePath value="$PROJECT_DIR$/Lab01/Server/Server.go" />
       <method v="2" />
     </configuration>
+    <configuration name="integration_test.go (1)" type="GoTestRunConfiguration" factoryName="Go Test" temporary="true" nameIsGenerated="true">
+      <module name="Distributed_System" />
+      <working_directory value="$PROJECT_DIR$/Lab04" />
+      <kind value="FILE" />
+      <directory value="$PROJECT_DIR$" />
+      <filePath value="$PROJECT_DIR$/Lab04/integration_test.go" />
+      <framework value="gotest" />
+      <method v="2" />
+    </configuration>
+    <configuration name="integration_test.go" type="GoTestRunConfiguration" factoryName="Go Test" temporary="true" nameIsGenerated="true">
+      <module name="Distributed_System" />
+      <working_directory value="$PROJECT_DIR$/Lab05" />
+      <kind value="FILE" />
+      <directory value="$PROJECT_DIR$" />
+      <filePath value="$PROJECT_DIR$/Lab05/integration_test.go" />
+      <framework value="gotest" />
+      <method v="2" />
+    </configuration>
     <recent_temporary>
       <list>
+        <item itemvalue="Go Test.integration_test.go (1)" />
+        <item itemvalue="Go Test.integration_test.go" />
         <item itemvalue="Go Build.go build Server.go" />
       </list>
     </recent_temporary>
@@ -81,8 +108,8 @@
   <component name="SharedIndexes">
     <attachedChunks>
       <set>
-        <option value="bundled-gosdk-d297c17c1fbd-d82d926f19eb-org.jetbrains.plugins.go.sharedIndexes.bundled-GO-243.22562.186" />
-        <option value="bundled-js-predefined-d6986cc7102b-deb605915726-JavaScript-GO-243.22562.186" />
+        <option value="bundled-gosdk-d297c17c1fbd-57c114c3cede-org.jetbrains.plugins.go.sharedIndexes.bundled-GO-243.25659.52" />
+        <option value="bundled-js-predefined-d6986cc7102b-1632447f56bf-JavaScript-GO-243.25659.52" />
       </set>
     </attachedChunks>
   </component>
