+++ Lab04/Shared/consistent_hash.go
