+++ Lab04/Shared/types.go
