+++ Lab04/go.mod
@@ -1,0 +1,3 @@
+module Lab04
+
+go 1.21
