--- Lab05/go.mod
@@ -1,3 +1,0 @@
-module Lab05
-
-go 1.21
